CREATE INDEX support through Calcite code path
Also moved DDL utils to src/frontend/org/voltdb/calciteutils directory.
