+++ src/frontend/org/voltdb/expressions/AbstractExpression.java
@@ -1402,7 +1402,6 @@ public abstract class AbstractExpression implements JSONString, Cloneable {
      * This function will recursively find any function expression with ID functionId.
      * If found, return true. Otherwise, return false.
      *
-     * @param expr
      * @param functionId
      * @return
      */
