+++ src/frontend/org/voltdb/expressions/VectorValueExpression.java
@@ -17,7 +17,11 @@
 
 package org.voltdb.expressions;
 
+import org.voltdb.VoltType;
 import org.voltdb.types.ExpressionType;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Represents a vector of expression trees.
@@ -27,6 +31,12 @@ public class VectorValueExpression extends AbstractExpression {
 
     public VectorValueExpression() {
         super(ExpressionType.VALUE_VECTOR);
+    }
+
+    public VectorValueExpression(List<AbstractExpression> args) {
+        this();
+        assert args != null : "Args for VectorValueExpression cannot be null";
+        setArgs(args);
     }
 
     @Override
