+++ tests/frontend/org/voltdb/newplanner/TestCalciteLogicalRules.java
@@ -26,7 +26,7 @@ package org.voltdb.newplanner;
 import org.apache.calcite.plan.RelOptUtil;
 import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.rel.RelRoot;
-import org.voltdb.catalog.org.voltdb.calciteadaptor.CatalogAdapter;
+import org.voltdb.calciteadapter.CatalogAdapter;
 import org.voltdb.newplanner.rules.PlannerPhase;
 import org.voltdb.types.CalcitePlannerType;
